Move driver-connection to lib
